Fix bug encountered when routing to different view after link created
The ref callback on the Link component would get called on unmount when
the page needed to switch to preferences. However, the element wouldn't
be passed through or something, so it was erroring. Now keeping track on
whether or not it has been loaded already to knoe whether or not to try
and highlight the text.